rename opencdems server package to opencdms api
