+++ pyproject.toml
@@ -1,11 +1,11 @@
 [tool.poetry]
-name = "opencdms-server"
+name = "opencdms-api"
 version = "0.1.0"
 description = ""
 authors = ["Your Name <you@example.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 fastapi = "^0.70.0"
 uvicorn = "^0.15.0"
 SQLAlchemy = "^1.4.26"
