+++ src/opencdms_api/db.py
@@ -1,5 +1,5 @@
 from contextlib import contextmanager
-from opencdms_server.config import settings
+from opencdms_api.config import settings
 from sqlalchemy.orm import sessionmaker, scoped_session, Session
 from sqlalchemy import create_engine
 from sqlalchemy.engine import Engine
