+++ src/opencdms_api/router.py
@@ -4,13 +4,13 @@ from fastapi import APIRouter, Depends
 from fastapi.exceptions import HTTPException
 from sqlalchemy.orm.session import Session
 from passlib.hash import django_pbkdf2_sha256 as handler
-from opencdms_server import deps, models
-from opencdms_server.schema import (
+from opencdms_api import deps, models
+from opencdms_api.schema import (
     UserCreateSchema,
     AuthenticationSchema,
     TokenSchema,
 )
-from opencdms_server.config import settings
+from opencdms_api.config import settings
 from jose import jwt
 
 router = APIRouter()
